# dist/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Hyperledger Explorer REST API Swagger",
        "description": "Rest API for fabric",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hyperledger Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "https://localhost:8080"
        }
    ],
...
    "paths": {
        "/api/channels": {
            "get": {
                "tags": ["Channel"],
                "summary": "Query to fetch channels",
                "operationId": "fetchChannel",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/channels/info": {
            "get": {
                "tags": ["Channel"],
                "summary": "Query to fetch channels info",
                "operationId": "fetchChannelInfo",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/curChannel": {
            "get": {
                "tags": ["Channel"],
                "summary": "Query to get Current Channel",
                "operationId": "getCurrentChannel",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/changeChannel/{channel_genesis_hash}": {
            "get": {
                "tags": ["Channel"],
                "summary": "Query to Change the Channel",
                "operationId": "getChangeChannel",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Query to Change the Channel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/block/{channel_genesis_hash}/{number}": {
            "get": {
                "tags": ["Blocks"],
                "summary": "Query to fetch block details",
                "operationId": "getBlocks",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide channel hash value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Please provide number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/txByOrg/{channel_genesis_hash}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to fetch transaction by Org",
                "operationId": "getTransByOrg",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please give the Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/fetchDataByTxnId/{channel_genesis_hash}/{txnId}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to fetch transaction by txnId",
                "operationId": "fetchDataByTxnId",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "txnId",
                        "in": "path",
                        "description": "Please provide transaction id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/fetchDataByBlockNo/{channel_genesis_hash}/{blockNo}": {
            "get": {
                "tags": ["Blocks"],
                "summary": "Query to fetch block details",
                "operationId": "fetchDataByBlockNo",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide channel hash value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "blockNo",
                        "in": "path",
                        "description": "Please provide blockNo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/fetchDataByBlockRange/{channel_genesis_hash}/{startBlockNo}/{endBlockNo}": {
            "get": {
                "tags": ["Blocks"],
                "summary": "Query to fetch block details",
                "operationId": "fetchDataByBlockNo",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide channel hash value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "startBlockNo",
                        "in": "path",
                        "description": "Please provide startBlockNo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "endBlockNo",
                        "in": "path",
                        "description": "Please provide endBlockNo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/metadata/{channel_genesis_hash}/{chaincode}": {
            "get": {
                "tags": ["Chaincodes"],
                "summary": "Query to fetch Chaincode metadata",
                "operationId": "getContractMetadata",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide channel hash value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "chaincode",
                        "in": "path",
                        "description": "Please provide chaincode",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/fetchBlockByTxId/{channel_genesis_hash}/{txnId}": {
            "get": {
                "tags": ["Blocks"],
                "summary": "Query to fetch block details",
                "operationId": "fetchBlockByTxId",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide channel hash value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "txnId",
                        "in": "path",
                        "description": "Please provide transaction id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/fetchBlockByHash/{channel_genesis_hash}/{block_hash}": {
            "get": {
                "tags": ["Blocks"],
                "summary": "Query to fetch block details",
                "operationId": "fetchBlockByHash",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide channel hash value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "block_hash",
                        "in": "path",
                        "description": "Please provide block hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/fetchEndorsersCommitter/{channel_genesis_hash}": {
            "get": {
                "tags": ["Channel"],
                "summary": "Query to fetch all Endorsers and Committers of channel",
                "operationId": "fetchEndorsersCommitter",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please provide channel hash value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}